requirements.txt
asgiref==3.8.1
crispy-bootstrap5==0.7
Django==5.1.1
django-crispy-forms==2.0
django-debug-toolbar==4.4.6
django-tables2==2.7.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
plotly==5.24.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
tzdata==2024.1
psycopg2-binary==2.9.3
gunicorn
environs[django]==9.5.0
django-allauth==65.2.0 
whitenoise==6.1.0
black